Added limitation to all transaction option and placed the limit to min:1.00$ and max of 10000$

See merge request smartpay-pod/smartpay!248

+++ backend/src/main/java/com/fdmgroup/SmartPay_BackEnd/services/payee/RecurringPaymentProcessorImpl.java
@@ -5,7 +5,6 @@ import java.time.ZoneOffset;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import com.fdmgroup.SmartPay_BackEnd.Utility.RecurringPaymentStatus;
 import org.springframework.stereotype.Service;
 
 import com.fdmgroup.SmartPay_BackEnd.Utility.TransactionExecutor;
@@ -53,7 +52,7 @@ public class RecurringPaymentProcessorImpl implements RecurringPaymentProcessor 
             return;
         }
 
-        walletService.debitRecurringPayment(payment.getOwner().getId(), payment.getAmount(),
+        walletService.debitRecurringPayment(payment.getOwner().getId(), payment.getAmount().doubleValue(),
                 payment.getPayeeName(), invocationDate);
         payment.setLastProcessedDate(invocationDate);
         payment.setDate(nextDateAfter(payment.getDate(), payment.getSchedule(), invocationDate));
